+++ pkg/api/update.go
@@ -6,7 +6,7 @@ import (
 	"reflect"
 	"strconv"
 
-	"github.com/Azure/open-service-broker-azure/pkg/async/model"
+	"github.com/Azure/open-service-broker-azure/pkg/async"
 	"github.com/Azure/open-service-broker-azure/pkg/service"
 	log "github.com/Sirupsen/logrus"
 	"github.com/gorilla/mux"
@@ -270,7 +270,7 @@ func (s *server) update(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	task := model.NewTask(
+	task := async.NewTask(
 		"updateStep",
 		map[string]string{
 			"stepName":   firstStepName,
