+++ pkg/async/fake/engine.go
@@ -3,31 +3,35 @@ package fake
 import (
 	"context"
 
-	"github.com/Azure/open-service-broker-azure/pkg/async/model"
+	"github.com/Azure/open-service-broker-azure/pkg/async"
 )
+
+// RunFn describes a function used to provide pluggable runtime behavior to
+// the fake async engine
+type RunFn func(context.Context) error
 
 // Engine is a fake implementation of async.Engine used for testing
 type Engine struct {
-	SubmittedTasks map[string]model.Task
+	SubmittedTasks map[string]async.Task
 	RunBehavior    RunFn
 }
 
 // NewEngine returns a new, fake implementation of async.Engine used for testing
 func NewEngine() *Engine {
 	return &Engine{
-		SubmittedTasks: make(map[string]model.Task),
+		SubmittedTasks: make(map[string]async.Task),
 		RunBehavior:    defaultEngineRunBehavior,
 	}
 }
 
 // RegisterJob registers a new Job with the async engine
-func (e *Engine) RegisterJob(name string, fn model.JobFn) error {
+func (e *Engine) RegisterJob(name string, fn async.JobFn) error {
 	return nil
 }
 
 // SubmitTask submits an idempotent task to the async engine for reliable,
 // asynchronous completion
-func (e *Engine) SubmitTask(task model.Task) error {
+func (e *Engine) SubmitTask(task async.Task) error {
 	e.SubmittedTasks[task.GetID()] = task
 	return nil
 }
