+++ pkg/async/redis/cleaner.go
@@ -1,4 +1,4 @@
-package async
+package redis
 
 import (
 	"context"
