+++ pkg/async/redis/cleaner_test.go
@@ -1,4 +1,4 @@
-package async
+package redis
 
 import (
 	"context"
