+++ pkg/async/redis/common.go
@@ -1,4 +1,4 @@
-package async
+package redis
 
 import "fmt"
