+++ pkg/async/task_test.go
@@ -1,4 +1,4 @@
-package model
+package async
 
 import (
 	"fmt"
