+++ pkg/broker/provision.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/Azure/open-service-broker-azure/pkg/async/model"
+	"github.com/Azure/open-service-broker-azure/pkg/async"
 	"github.com/Azure/open-service-broker-azure/pkg/service"
 	log "github.com/Sirupsen/logrus"
 )
@@ -130,7 +130,7 @@ func (b *broker) doProvisionStep(
 				"error persisting instance",
 			)
 		}
-		task := model.NewTask(
+		task := async.NewTask(
 			"provisionStep",
 			map[string]string{
 				"stepName":   nextStepName,
